+++ pkg/cmd/getMeasurementCollectionCmd.auto.go
@@ -40,8 +40,8 @@ Get a list of measurements
 	cmd.Flags().String("dateFrom", "", "Start date or date and time of measurement occurrence.")
 	cmd.Flags().String("dateTo", "", "End date or date and time of measurement occurrence.")
 	cmd.Flags().Bool("revert", false, "Return the newest instead of the oldest measurements. Must be used with dateFrom and dateTo parameters")
-	cmd.Flags().Bool("csv", false, "Results will be displayed in csv format. Note: -IncludeAll, is not supported when using using this parameter")
-	cmd.Flags().Bool("excel", false, "Results will be displayed in Excel format Note: -IncludeAll, is not supported when using using this parameter")
+	cmd.Flags().Bool("csvFormat", false, "Results will be displayed in csv format. Note: -IncludeAll, is not supported when using using this parameter")
+	cmd.Flags().Bool("excelFormat", false, "Results will be displayed in Excel format Note: -IncludeAll, is not supported when using using this parameter")
 	cmd.Flags().String("unit", "", "Every measurement fragment which contains 'unit' property will be transformed to use required system of units.")
 
 	flags.WithOptions(
@@ -99,8 +99,8 @@ func (n *GetMeasurementCollectionCmd) RunE(cmd *cobra.Command, args []string) er
 		cmd,
 		headers,
 		inputIterators,
-		flags.WithBoolValue("csv", "Accept", "text/csv"),
-		flags.WithBoolValue("excel", "Accept", "application/vnd.ms-excel"),
+		flags.WithBoolValue("csvFormat", "Accept", "text/csv"),
+		flags.WithBoolValue("excelFormat", "Accept", "application/vnd.ms-excel"),
 		flags.WithStringValue("unit", "X-Cumulocity-System-Of-Units"),
 	)
 	if err != nil {